# openapi.yaml
openapi: 3.0.3
info:
  title: TalkBlog Back API
  version: 1.0.0
  description: OpenAPI spec for TalkBlog backend services
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/admin/users/{userId}/status:
    patch:
      tags:
        - Admin Users
      summary: Update user status
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: Status updated
  /api/comments:
    post:
      tags:
        - Comments
      summary: Create a comment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
          multipart/form-data:
            schema:
              type: object
              required:
                - blogId
                - content
              properties:
                blogId:
                  type: integer
                  format: int64
                content:
                  type: string
                  description: 评论内容，支持 Emoji，按 Unicode 码点计数
                parentId:
                  type: integer
                  format: int64
                replyToUserId:
                  type: integer
                  format: int64
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - Comments
      summary: Get comments for a blog
      parameters:
        - name: blogId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: parentId
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Comment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateUserStatusRequest:
      type: object
      properties:
        status:
          type: integer
    CreateCommentRequest:
      type: object
      required:
        - blogId
        - content
      properties:
        blogId:
          type: integer
          format: int64
        content:
          type: string
          maxLength: 500
          description: 评论内容，支持 Emoji，按实际 Unicode 码点计数（最多 500 个码点）
        parentId:
          type: integer
          format: int64
        replyToUserId:
          type: integer
          format: int64
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...